tests: organize subscriptions bp and add basic tests
